# Conflicts:
#	MyPortal.Auth/Interfaces/IPermissionService.cs
#	MyPortal.Core/Entities/Role.cs
#	MyPortal.Migrations/MyPortal.Migrations.csproj
#	MyPortal.Tests/ServiceTests/UserServiceTests.cs
#	MyPortal.WebApi/SystemSeedVerifier.cs

+++ MyPortal.Services/System/UserService.cs
@@ -174,6 +174,11 @@ public class UserService : BaseService, IUserService
             throw new NotFoundException("User not found.");
         }
 
+        if (user.IsSystem)
+        {
+            throw new SystemEntityException("System users cannot be modified.");
+        }
+
         bool userDisabled = user.IsEnabled && !model.IsEnabled;
         bool userTypeChanged = user.UserType != model.UserType;
 
@@ -223,6 +228,11 @@ public class UserService : BaseService, IUserService
         if (user == null)
         {
             throw new NotFoundException("User not found.");
+        }
+
+        if (user.IsSystem)
+        {
+            throw new SystemEntityException("System users cannot be deleted.");
         }
 
         var result = await _userManager.DeleteAsync(user);
